Removes IonElement trait, ElementRef impl
`IonElement` provides a trait abstraction over two flavors of
`Element` API:

1. The 'borrowed' API, where each `ElementRef` has a lifetime
   tied to the lifetime of the source data.

2. The 'owned' API, where each `Element` is self-contained, with
   no connection to the source data.

While the borrowed API should be lighter weight, the current
implementation performs materializes nearly the same amount of data
as the owned API. Collections (`list`, `sexp`, `struct`) are
heap-allocated; only `string`s and `lob`s avoid allocation.

The `IonElement` trait was designed before GATs were stabilized
and had some rough edges caused by the inherent difficulty in
abstracting over lifetime constraints in traits in the absence of GATs.

In practice, users that need an `Element` API are exclusively using
the `owned` variant.

In light of these observations and in anticipation of the upcoming v1.0
release, this PR removes both the `IonElement` trait and the `borrowed`
implementation, leaving only the owned `Element` API. These constructs
can be reintroduced in the future if and when they are needed.