Introduces `Str` type
This PR introduces a new `Str` type that replaces all usages of
Rust's owned `String` type in the `Reader` and `Element` APIs.

For the time being, the `Str` type is a simple opaque wrapper
around `String` with trait implementations that allow it to be
used in much the same way. Having an oqaque wrapper will allow
us to swap out the implementation later without a breaking
change, enabling us to provide stack-allocated small strings,
shared strings (interning), etc as desired.

The name `Str` was selected to:
1. Avoid confusing collisions with `String` and `&str`.
2. Use the same type name as Ion itself (vs, say, `Text`)

We can add more trait implementations and method passthroughs
as needed.