Makes `List` and `SExp` thin wrappers around `Sequence`
See issue #500 for discussion.

This PR removes the `IonSequence` trait and reintroduces a
concrete `Sequence` type. It also converts `List` and `SExp`
into tuple structs that wrap `Sequence`.

`ListBuilder` and `SExpBuilder` have also been removed in favor
of a single `SequenceBuilder` implementation with `build_list()`
and `build_sexp()` methods.

`List` and `SExp` each implement `Deref<Target=Sequence>`, allowing
them both to expose the functionality of their underlying `Sequence`
of elements transparently.

This eliminates a fair amount of boilerplate code that was needed
to provide functionality for both types. It also enables unified
handling for `List` and `SExp` in match expressions in situations
where the author does not care about the Ion type distinction.