Fixes rustdoc to build on all private items
Adds a CI step to run `cargo doc` for all features and private items
going forward which will require all users to keep their rustdoc clean.

For the parts of the docs I cleaned up, I tried to put more code
references in an inline code fence--this is required for types with `<`
or `>`, but also the style I think we should use.  I did not go through
all the code to clean this up.