Adds Lazy Evaluation Support
Adds `thunk::Thunk` as an experimental API for modeling lazy/eager
values that are fallible (and thus evaluate to an `IonResult`.
The API supports controlling how the value is materialized from a
closure and supports moving values into and out of it.

The `Thunk` models three distinct states:

* ***Deferred***: value is backed by a closure.
* ***Materialized***: value is stored directly in the `Thunk`.
* ***Error***: there was some error evaluating the *deferred* value.
  This state can be thought of as a deferred value that can *never* be
  evaluated.

This also backs this API with a Cargo feature called "experimental" that
we can use for other such APIs.

resolves: #519, #521