Adds CI matrix for crate features
This change makes it so we build our crate with all the various
interesting permutations:

* *default* - build with no features set.
* `--all-features` - build with all the bells and whistles.
* `--features X` where `X` is:
  * `ion-hash`
  * `experimental`
  * `ion-hash,experimental`

When building/testing with these permutations, it uncovered a bug in the
`ion-hash` doc tests where the test code depends on an optional feature.
This has been conditionally compiled out, preserving the text of the
doc comment (only runs when the appropriate feature flag is enabled).

Also updates the build of API docs to error out on leaking internal
references.  This change cleans up such documentation comments.