Fixes clippy warnings
Made clippy required for CI--we tend to keep the source tree clean, but
the clippy warnings are easy to miss and cannot be seen on PRs from user
forks.  by adding `-Dwarnings` to the clippy CI step, we make sure this
breaks the build--we can always add annotations to ignore suggestions
explicitly and/or we can back this out if clippy is too annoying.

Also adds `--workspace` to the clippy, as it doesn't seem to check
`ion-hash` by default without it.