Updates README (#537)
The original text suggested using `--all-targets` for building.  This
replaces that with `--all-features` as the suggestion for both building
and testing.  Removing `--all-targets`, especially for tests, is good
because doc tests don't get built/run with that parameter.

Adds a clean/build everything script to make it easier to run most of what
CI does.  Adds this to the README as well.