Adds TokenStreamReader
`TokenStreamReader` adapts `TokenStream` into a user-level `Reader`.

This also changes the support types for `ElementReader` to work with
`IonReader` trait objects (adding `?Sized` to the type constraints).

The `Element` round-trip tests are also adapted to work with _any_
`IonReader` as a boxed trait object, which allows us to slot in
`TokenStreamReader` over `ReaderTokenStream` over a user `Reader`
instance, exercising both adapters in this framework.

This also consolidates the round-trip skip-lists into constants so
there is less copy/paste between the now three instances of the
`ElementApi` trait.

resolves #520