Adds experimental tokens module (#540)
This is the basis for `TokenStream` and the implementations bridging to
and from `IonReader`.  The `TokenStream` is a minimal interface where
(potentially) lazy `Token` instances are returned instead of `read_XXX`
methods.  This allows us to compose streams of Ion values more easily
because we now have an iterator-like interface on which we can model
the computation of an Ion data stream.  Unlike iterator, `TokenStream`
has as its input an `Instruction` which allows a user to control
"skipping" to the end of a container, which is the analog of `step_in`
and `step_out` in the `IonReader` trait.

`Token` values have annotations and a field name (which could be
lazily evaluated).  They contain a `Content` which is either the
start/end of a container, a scalar value, typed null, or the end of a
stream.  These values rely on the `thunk` module to provide deferring
the computation of the value.

The `token` module also provides various specific subdomains of
`IonType` in `ContainerType` and `ScalarType` to avoid contexts
where inapplicable `IonType` are possible (e.g., in a start/end
container token).

Also updates `rstest` and adds `rstest_reuse` for reusing table inputs
for unit tests.

Prerequisite for #520.

Co-authored-by: Zack Slayton <zack.slayton@gmail.com>