Replaces `Vec<Element>` with `Sequence`
APIs that returned `Vec<Element>` now return a `Sequence` instead.
This allows us to change the storage type without a breaking change.
It also means that you can trivially construct a `List` or `SExp`
from the output of `Element::read_all`.

This also renames `ElementsIterator` to `SequenceIterator` to avoid
confusion with the similarly named `ElementIterator`.