Replaces `Vec<Element>` with `Sequence` (#572)
* Replaces `Vec<Element>` with `Sequence`

APIs that returned `Vec<Element>` now return a `Sequence` instead.
This allows us to change the storage type without a breaking change.
It also means that you can trivially construct a `List` or `SExp`
from the output of `Element::read_all`.

This also renames `ElementsIterator` to `SequenceIterator` to avoid
confusion with the similarly named `ElementIterator`.

* ion_seq! can also accept comma-delimited args

* ion_list! and ion_struct! now accept trailing commas

---------

Co-authored-by: Zack Slayton <zslayton@amazon.com>