Makes `IonError` variants wrap opaque types (#584)
This PR refactors the `IonError` enum variants to each wrap an opaque
struct type. The wrapped types each offer methods and trait implementations
that we can expand upon in the future without breaking changes.

It also creates a new crate-visible trait, `IonFailure`, that adds extension
methods for constructing the opaque types and wrapping them in either
an `IonError` or an `IonResult` unambiguously.