Removes the crate's dependency on `BigInt`, `BigUint` (#767)
* Removes the dependency on BigInt, BigUint

This PR limits the range of integers that can be read and written
to those that can be represented by the `i128` type.

This reduces the size of `Int`, `UInt`, `Decimal`, and `Timestamp`,
`RawValueRef`, and `ValueRef`. It also allows those types to
implement `Copy`.

It also allows us to remove several custom methods to compare
across mixed `i64`/`BigInt` and `u64`/`BigUint` representations
for equality and ordering.