Adds support for qualified refs to `$ion`, `$ion_encoding`
* Adds support for `module_name::encoding_name::parameter` syntax
* Adds support for `(module_name::macro_name ...)` syntax

The only modules currently supported are `$ion` (the system module)
and `$ion_encoding` the active module.

This patch also starts uses `Rc` to store macro references in the body
of a compiled macro. This allows the macro to continue referring to
its dependent even if the dependent is no longer directly addressable.