Overhauls system macro table bootstrapping (#872)
This patch modifies `EncodingContext` to have an empty macro table when
constructed for use in v1.0 readers.

This allows the Ion 1.0 reader to be used in the process of bootstrapping
the Ion 1.1 system macro table, which was previously not possible due to
the (now eliminated) circular dependency between macros and readers.

System macros that can be expressed in TDL are now constructed from
a template definition encoded in Ion 1.0. As a result, the system macro
table initialization logic is now substantially easier to read and
more straightforward to modify.