Split `WriterContext` into separate macros, symbols fields
Because new symbols can be added to the symbol table at any depth,
each value writer needs to hold a mutable reference to the symbol table.

Because the ApplicationEExpWriter needs to hold information about the
macro being invoked, it needs to hold a shared reference to a macro
in the macro table.

These two requirements conflict if the symbol and macro tables both
reside in the same structure.

This commit splits the unified `WriterContext` into separate references.